service/requirements.txt
fastapi
uvicorn
google-genai
supabase
